+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
-import "./i18n";
+import "./config/locales/i18n";
 
 import Routes from "./routes";
 import Footer from "./components/modules/footer";
@@ -9,7 +9,7 @@ import Header from "./components/modules/header";
 
 import "./styles/main.scss";
 
-ReactDOM.render(
+const App = () => (
   <React.StrictMode>
     <BrowserRouter>
       <Header />
@@ -18,6 +18,7 @@ ReactDOM.render(
       </div>
       <Footer />
     </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
+
+ReactDOM.render(<App />, document.getElementById("root"));
